Relocate check in adequate package
