+++ plugins/java-custom-rules/src/main/java/org/sonar/samples/java/MyJavaFileCheckRegistrar.java
@@ -7,6 +7,7 @@ import org.sonar.plugins.java.api.JavaCheck;
 import org.sonar.samples.java.checks.AvoidAnnotationCheck;
 import org.sonar.samples.java.checks.AvoidBrandInMethodNamesCheck;
 import org.sonar.samples.java.checks.AvoidMethodDeclarationCheck;
+import org.sonar.samples.java.checks.MyCustomSubscriptionRule;
 import org.sonar.samples.java.checks.SecurityAnnotationMandatoryCheck;
 
 /**
