+++ plugins/java-custom-rules/src/main/java/org/sonar/samples/java/checks/MyCustomSubscriptionRule.java
@@ -1,4 +1,4 @@
-package org.sonar.samples.java;
+package org.sonar.samples.java.checks;
 
 import java.util.List;
